STUDY-07 | Adding new record RequestWordDefinitions to better deserialise the incoming JSON to an easy-to-understand object. Updating WordService to remove WordDefinition logic to a new WordDefinitionService. Adding WordDtoService to prevent dependency loops between CategoryService and WordService
