+++ src/main/java/com/japanese/study_app/request/RequestWordDefinitions.java
@@ -1,0 +1,9 @@
+package com.japanese.study_app.request;
+
+import java.util.Set;
+
+public record RequestWordDefinitions(
+        Set<String> english,
+        Set<String> japanese
+) {
+}
