+++ src/main/java/com/japanese/study_app/request/UpdateWordRequest.java
@@ -12,7 +12,7 @@ public record UpdateWordRequest(
         Set<EnglishWord> englishWord,
         String hiragana,
         Set<Map<String, String>> exampleSentence,
-        Map<String, Set<String>> definitions,
+        RequestWordDefinitions definitions,
         Set<Category> category
 ) {
